separate responsibilities of tests
